Move implementation of bpf.h to bpf.cc
